Nav bar view and costodian
